backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pyjwt
passlib[bcrypt]
python-dotenv
openai
httpx
sentence-transformers
chromadb
PyPDF2
python-docx
Pillow
supabase==1.0.3
langgraph
langchain-core
langchain-openai
tavily-python
elevenlabs
python-multipart
edge-tts
